+++ js/Screens/levels/game-view.js
@@ -1,7 +1,7 @@
 import {resize} from "../../data/resize";
 import AbstractView from "../Abstract";
 import {curStats} from "../../current-stats";
-import {BackButton} from "../../render/app-screens";
+import BackButton from "../../constrollers/back-button";
 import GameTimer from "../header/time";
 import GameLives from "../header/live";
